refactor specreader utils into separate package (#7783)
